Hello!

Found your library very useful and I found a very simple optimization that made it 70% faster. I also added a very simple benchmark to test that change out.

```go
benchmark                old ns/op     new ns/op     delta
BenchmarkUnmarshal-8     8066          2293          -71.57%

benchmark                old allocs     new allocs     delta
BenchmarkUnmarshal-8     29             11             -62.07%

benchmark                old bytes     new bytes     delta
BenchmarkUnmarshal-8     1968          576           -70.73%
```

I didn't think it would be necessary to open a issue and decided to open a PR directly. Hope you enjoy it!

Thanks