@TomWright 2018-10-07
Well spotted! I'm glad you find the package useful and thanks a lot for contributing.